Added ready trigger for BasePage
BasePage now accepts a page component type as opposed to an instantiated element of said component. BasePage will explicitly make an element for the page component.

BasePage now relies on a "ready" function to be invoked in order to change the loading state. The ready function is passed as a prop to the page component, where this child page component must invoke this ready function in order for the loading screen to disappear.